Renamed PermScopeConstants to Constants. Extracted aux classes to new file
