+++ PermissionScope/Extensions.swift
